feat: add generator TOC functionality
