--- data/blog/releases/README.md
@@ -1,1 +1,0 @@
-RELEASES ARTICLE
