--- data/blog/tutorials/README.md
@@ -1,1 +1,0 @@
-TUTORIAL ARTICLE
