+++ data/blog/tutorials/optimize-images-for-web-performance.mdx
@@ -92,8 +92,6 @@ relatedPosts:
 blog: true
 ---
 
-## Takeaways
-
 <Checklist
   items={[
     "Serve images in next-gen formats: Use AVIF and WebP, with JPG as fallback",
